+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Limit the size of propcombined groups to avoid hitting vertex limits.
 * Prevent automatically packing models specified only as Hammer previews in various entities.
 * Fix propcombine sometimes removing collisions entirely from component props.
+* Add an option to allow the sources for compiled models to be preserved.
 
 --------------------